DAY9-LangChain-Memory-Tools/Memory/requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-core
google-generativeai
python-dotenv
